STYLE: Move definition of `ModelConfiguration` into its own header file
Moved `ModelConfiguration` from inside `ImpactImageToImageMetric<TFixedImage, TMovingImage>` into its own header file, `itkImpactModelConfiguration.h`, and renamed it to `ImpactModelConfiguration`.

Aims to simplify the design, by expressing that there is just one ModelConfiguration type for the IMPACT metric, independent of the types `TFixedImage` and `TMovingImage`.
